requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
glob2==0.7
importlib-metadata==0.19
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
parse==1.12.0
parse-type==0.5.2
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytest-bdd==3.1.1
pytest-html==1.22.0
pytest-metadata==1.8.0
rope==0.14.0
ruamel.yaml==0.16.1
ruamel.yaml.clib==0.1.2
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
